crawler/requirements.txt
# Core SDKs
google-generativeai==0.8.5
google-ai-generativelanguage>=0.6.15,<1
firecrawl-py==4.5.0

# LangChain integrations
langchain==0.3.27
langchain-google-genai==2.0.10

# Env & utilities
python-dotenv==1.0.1
tqdm==4.67.1
pydantic==2.12.3

# Frontend
streamlit==1.50.0
